Amended dockerfile and started to implement the cli interface

+++ get-nitro-k8s-pipe.sh
@@ -22,7 +22,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-mkdir .nitro && cd .nitro
-curl -fsSL -o nitro_k8s_docker_pipe.sh https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/nitro-k8s-docker-pipe.sh
-curl -fsSL -o nitro_k8s_pipe.sh https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/nitro-k8s-pipe.sh
-curl -fsSL -o Dockerfile https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/Dockerfile
+{ [ -d .nitro ] || mkdir .nitro; cd .nitro; }
+[ -f nitro_k8s_docker_pipe.sh ] || curl -fsSL -o nitro_k8s_docker_pipe.sh https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/nitro-k8s-docker-pipe.sh
+[ -f nitro_k8s_pipe.sh ] || curl -fsSL -o nitro_k8s_pipe.sh https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/nitro-k8s-pipe.sh
+[ -f Dockerfile ] || curl -fsSL -o Dockerfile https://raw.githubusercontent.com/NitroAgility/nitro-k8s-pipe/master/Dockerfile
